+++ extensions/in-memory/contractdefinition-store-memory/src/main/java/org/eclipse/dataspaceconnector/contractdefinition/store/memory/InMemoryContractDefinitionStore.java
@@ -14,12 +14,10 @@
 
 package org.eclipse.dataspaceconnector.contractdefinition.store.memory;
 
-import org.eclipse.dataspaceconnector.common.reflection.ReflectionUtil;
 import org.eclipse.dataspaceconnector.spi.contract.offer.store.ContractDefinitionStore;
-import org.eclipse.dataspaceconnector.spi.query.BaseCriterionToPredicateConverter;
-import org.eclipse.dataspaceconnector.spi.query.Criterion;
+import org.eclipse.dataspaceconnector.spi.query.QueryResolver;
 import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
-import org.eclipse.dataspaceconnector.spi.query.SortOrder;
+import org.eclipse.dataspaceconnector.spi.query.ReflectionBasedQueryResolver;
 import org.eclipse.dataspaceconnector.spi.types.domain.contract.offer.ContractDefinition;
 import org.jetbrains.annotations.NotNull;
 
@@ -27,16 +25,14 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.function.Predicate;
 import java.util.stream.Stream;
-
-import static org.eclipse.dataspaceconnector.common.reflection.ReflectionUtil.propertyComparator;
 
 /**
  * The default store implementation used when no extension is configured in a runtime. {@link ContractDefinition}s are stored ephemerally in memory.
  */
 public class InMemoryContractDefinitionStore implements ContractDefinitionStore {
     private final Map<String, ContractDefinition> cache = new ConcurrentHashMap<>();
+    private final QueryResolver<ContractDefinition> queryResolver = new ReflectionBasedQueryResolver<>(ContractDefinition.class);
 
     @Override
     public @NotNull Collection<ContractDefinition> findAll() {
@@ -45,25 +41,8 @@ public class InMemoryContractDefinitionStore implements ContractDefinitionStore 
 
     @Override
     public @NotNull Stream<ContractDefinition> findAll(QuerySpec spec) {
-        var stream = cache.values().stream();
+        return queryResolver.query(cache.values().stream(), spec);
 
-        //filter
-        var andPredicate = spec.getFilterExpression().stream().map(this::toPredicate).reduce(x -> true, Predicate::and);
-        stream = stream.filter(andPredicate);
-
-        //sort
-        var sortField = spec.getSortField();
-
-        if (sortField != null) {
-            if (ReflectionUtil.getFieldRecursive(ContractDefinition.class, sortField) == null) {
-                return Stream.empty();
-            }
-            var comparator = propertyComparator(spec.getSortOrder() == SortOrder.ASC, sortField);
-            stream = stream.sorted(comparator);
-        }
-
-        // limit
-        return stream.skip(spec.getOffset()).limit(spec.getLimit());
     }
 
     @Override
@@ -91,15 +70,4 @@ public class InMemoryContractDefinitionStore implements ContractDefinitionStore 
         // no-op
     }
 
-    private Predicate<ContractDefinition> toPredicate(Criterion criterion) {
-        return new ContractDefinitionPredicateConverter().convert(criterion);
-    }
-
-
-    private static class ContractDefinitionPredicateConverter extends BaseCriterionToPredicateConverter<ContractDefinition> {
-        @Override
-        protected <R> R property(String key, Object object) {
-            return ReflectionUtil.getFieldValueSilent(key, object);
-        }
-    }
 }