+++ extensions/in-memory/transfer-store-memory/src/main/java/org/eclipse/dataspaceconnector/transfer/store/memory/InMemoryTransferProcessStore.java
@@ -15,9 +15,9 @@
 package org.eclipse.dataspaceconnector.transfer.store.memory;
 
 import org.eclipse.dataspaceconnector.common.concurrency.LockManager;
-import org.eclipse.dataspaceconnector.spi.query.Criterion;
+import org.eclipse.dataspaceconnector.spi.query.QueryResolver;
 import org.eclipse.dataspaceconnector.spi.query.QuerySpec;
-import org.eclipse.dataspaceconnector.spi.query.SortOrder;
+import org.eclipse.dataspaceconnector.spi.query.ReflectionBasedQueryResolver;
 import org.eclipse.dataspaceconnector.spi.transfer.store.TransferProcessStore;
 import org.eclipse.dataspaceconnector.spi.types.domain.transfer.TransferProcess;
 import org.jetbrains.annotations.NotNull;
@@ -30,12 +30,10 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.toList;
-import static org.eclipse.dataspaceconnector.common.reflection.ReflectionUtil.propertyComparator;
 
 /**
  * An in-memory, threadsafe process store.
@@ -46,6 +44,7 @@ public class InMemoryTransferProcessStore implements TransferProcessStore {
     private final Map<String, TransferProcess> processesById = new HashMap<>();
     private final Map<String, TransferProcess> processesByExternalId = new HashMap<>();
     private final Map<Integer, List<TransferProcess>> stateCache = new HashMap<>();
+    private final QueryResolver<TransferProcess> queryResolver = new ReflectionBasedQueryResolver<>(TransferProcess.class);
 
     @Override
     public TransferProcess find(String id) {
@@ -126,27 +125,7 @@ public class InMemoryTransferProcessStore implements TransferProcessStore {
     public Stream<TransferProcess> findAll(QuerySpec querySpec) {
         return lockManager.readLock(() -> {
             Stream<TransferProcess> transferProcessStream = processesById.values().stream();
-            // filter
-            var andPredicate = querySpec.getFilterExpression().stream().map(this::toPredicate).reduce(x -> true, Predicate::and);
-            transferProcessStream = transferProcessStream.filter(andPredicate);
-
-            // sort
-            var sortField = querySpec.getSortField();
-
-            if (sortField != null) {
-                var comparator = propertyComparator(querySpec.getSortOrder() == SortOrder.ASC, sortField);
-                transferProcessStream = transferProcessStream.sorted(comparator);
-            }
-
-            //limit
-            transferProcessStream = transferProcessStream.skip(querySpec.getOffset()).limit(querySpec.getLimit());
-
-            return transferProcessStream;
+            return queryResolver.query(transferProcessStream, querySpec);
         });
     }
-
-    private Predicate<TransferProcess> toPredicate(Criterion criterion) {
-        return new TransferProcessPredicateConverter().convert(criterion);
-    }
-
 }